refactor(server/air quality): refactor utils into separate file
