+++ server/src/api/airQuality/airQuality.externalMappings.ts
@@ -1,9 +1,6 @@
 import type { AirQualitySchemaExternal } from './airQuality.externalSchema.js'
-import type {
-  AirQuality,
-  AlertThresholds,
-  PollutionTypes,
-} from './airQuality.types.js'
+import type { AirQuality, PollutionTypes } from './airQuality.types.js'
+import { determineAlertThreshold } from './airQuality.utils.js'
 
 const ALLOWED_EXTERNAL_POLLUTION_TYPES = ['PM10', 'PM25']
 
@@ -28,23 +25,3 @@ export const mapAirQualityDataToAirQuality = (
       }
       return accumulator
     }, {} as AirQuality)
-
-const determineAlertThresholdForPM10 = (value: number): AlertThresholds => {
-  if (value <= 20) return 'low'
-  if (value <= 50) return 'medium'
-  return 'high'
-}
-
-const determineAlertThresholdForPM25 = (value: number): AlertThresholds => {
-  if (value <= 10) return 'low'
-  if (value <= 25) return 'medium'
-  return 'high'
-}
-
-const determineAlertThreshold = (
-  name: PollutionTypes,
-  value: number,
-): AlertThresholds => {
-  if (name === 'PM10') return determineAlertThresholdForPM10(value)
-  if (name === 'PM25') return determineAlertThresholdForPM25(value)
-}
